refactor: привести код к единому стилю
