+++ README.md
@@ -1,9 +1,11 @@
 # Проект: Место
 
-Студент: Александр Дудин  
+Студенты: Александр Дудин и Мария Трифанова
 Наставники: Мария Шукшина, Алексей Гмитрон, Алексей Попков
 21-й поток профессии “Веб-разработчик плюс”  
 [Ссылка на опубликованный проект](https://adudin.github.io/mesto-project/)
+[Ссылка на репозиторий Марии](https://github.com/M0riaz/mesto-project)
+Электронная почта Марии: ulckiorra2010@yandex.ru
 
 ### Использованы технологии:
 * Adoptive & responsive design
@@ -14,6 +16,7 @@
 * Basic vanilla JS dynamic content adding/removing to/from DOM
 * Custom JS form validation
 * Webpack project assembling
+* OOP principles
 
 **Описание**
